Analysis_F_Proyect/requirements.txt
django
djangorestframework
djangorestframework_simplejwt
psycopg2
python-dotenv
gunicorn
dj-database-url